add redirect after submitting new template headers
