+++ app/createTemplate/[id]/page.js
@@ -1,7 +1,5 @@
-import { BackLink } from "../page"
+import Link from "next/link"
 import Template from "@/database/models/Templates"
-import CustomButton from "@/components/Buttons"
-import TemplateHeader from "@/components/TemplateHeader"
 
 export default async function NewTemplate({ params }) {
   const { id } = params
@@ -12,9 +10,9 @@ export default async function NewTemplate({ params }) {
 
     return (
         <div>
-            <BackLink href="/"> &lt; Back </BackLink>
-            <TemplateHeader name={template.name} focus={template.focus} />
-            <CustomButton textContent="Add Day" action="subtract" />
+            <Link href="/"> &lt; Back </Link>
+            <div>{template.name}</div>
+            <button type="button">Add Day</button>
         </div>
     )
 }
